chore(youtube-chan): extract channel cache key model/methods to file
