+++ internal/resolvers/youtube/helpers.go
@@ -31,33 +31,6 @@ func getYoutubeVideoIDFromURL2(url *url.URL) string {
 	return path.Base(url.Path)
 }
 
-func getChannelTypeFromString(channelType string) channelType  {
-	switch channelType {
-		case "c":
-			return CustomChannel
-		case "user":
-			return UserChannel
-		case "channel":
-			return IdentifierChannel
-	}
-
-	return InvalidChannel
-}
-
-func constructCacheKeyFromChannelId(id channelId) string {
-	return string(id.chanType) + ":" + id.id
-}
-
-func deconstructChannelIdFromCacheKey(cacheKey string) channelId  {
-	splitKey := strings.Split(cacheKey, ":")
-
-	if len(splitKey) < 2 {
-		return channelId{id: "", chanType: InvalidChannel}
-	}
-
-	return channelId{id: splitKey[1], chanType: getChannelTypeFromString(splitKey[0])}
-}
-
 func getYoutubeChannelIdFromURL(url *url.URL) channelId {
 	pattern, err := regexp.Compile(`(user|c(?:hannel)?)/([\w-]+)`)
 	if err != nil {
